Warning Fixed : CachedFileStream.cpp : Fixed many warnings
svn-id: r1120

+++ gemrb/plugins/Core/CachedFileStream.cpp
@@ -15,7 +15,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
  *
- * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/CachedFileStream.cpp,v 1.17 2003/11/29 10:33:20 balrog994 Exp $
+ * $Header: /data/gemrb/cvs2svn/gemrb/gemrb/gemrb/plugins/Core/CachedFileStream.cpp,v 1.18 2003/12/15 09:16:30 balrog994 Exp $
  *
  */
 
@@ -113,7 +113,7 @@ int CachedFileStream::Read(void * dest, int length)
 	if(Encrypted)
 		ReadDecrypted(dest, c);
 	Pos+=c;
-	return (int)c;
+	return c;
 }
 
 int CachedFileStream::Seek(int pos, int startpos)
